perf(board): one byte per square coordinate brings game.Move back to 16 bytes
The Promo field added for underpromotion replay grew game.Move from 32 to
40 bytes (two Sq of two ints plus a PieceType), and the search copies a
move into every move list, killer slot and history probe. board.Sq now
holds File and Rank as int8, so a Move is 16 bytes and its API is
unchanged. Coordinates never leave -2..9, so int8 arithmetic cannot
overflow; the padded cell index and the evaluation terms convert to int
before they compute. TestMoveFitsIn32Bytes guards the size (red at 40).

Seeded node count unchanged at 641868. Base d4ff76a against this commit,
BenchmarkChampionSearch -benchtime 5x, 5 alternated runs each under nice:
instructions retired 46.50G to 42.85G (-7.85%), ns/op 434.5 ms to
373.9 ms (-14.0%).

+++ board/board_test.go
@@ -6,11 +6,11 @@ func TestInitialSetupBackRank(t *testing.T) {
 	b := Initial()
 	order := []PieceType{Rook, Knight, Bishop, Queen, King, Bishop, Knight, Rook}
 	for file, pt := range order {
-		white, ok := b.PieceAt(Sq{file, 0})
+		white, ok := b.PieceAt(Sq{int8(file), 0})
 		if !ok || white.Color != White || white.Type != pt {
 			t.Errorf("file %d rank 0: got %+v ok=%v, want %v", file, white, ok, pt)
 		}
-		black, ok := b.PieceAt(Sq{file, 7})
+		black, ok := b.PieceAt(Sq{int8(file), 7})
 		if !ok || black.Color != Black || black.Type != pt {
 			t.Errorf("file %d rank 7: got %+v ok=%v, want %v", file, black, ok, pt)
 		}
@@ -19,7 +19,7 @@ func TestInitialSetupBackRank(t *testing.T) {
 
 func TestInitialSetupPawns(t *testing.T) {
 	b := Initial()
-	for file := 0; file < 8; file++ {
+	for file := int8(0); file < 8; file++ {
 		if p, ok := b.PieceAt(Sq{file, 1}); !ok || p.Color != White || p.Type != Pawn {
 			t.Errorf("file %d rank1: got %+v ok=%v", file, p, ok)
 		}
@@ -31,8 +31,8 @@ func TestInitialSetupPawns(t *testing.T) {
 
 func TestInitialSetupEmptyMiddle(t *testing.T) {
 	b := Initial()
-	for file := 0; file < 8; file++ {
-		for rank := 2; rank < 6; rank++ {
+	for file := int8(0); file < 8; file++ {
+		for rank := int8(2); rank < 6; rank++ {
 			if _, ok := b.PieceAt(Sq{file, rank}); ok {
 				t.Errorf("file %d rank %d should be empty", file, rank)
 			}
